Create Summing elements of an integer array
Given a list of integer data, you're tasked with summing the elements of the list and returning the sum. Fill the blanks in the Java code to create a method 'arraySum' that accepts an integer array 'arrayToSum' and returns the sum of all the elements. Also, fill the gaps to call this method in the main function with a created list of integers from 1 to 5.