Create The while and do-while loops
The length of the sequence
For its input, the program gets a sequence of non-negative integers; each integer is written on a separate line. The sequence ends with an integer 0; when the program reads 0, it should end its work and output the length of the sequence (not counting the final 0).

Don’t read numbers following the number 0.