Create Time constants
There are special types of classes called or which have only static fields and/or methods. Usually, such classes cannot be instantiated directly and they are only needed to group related static members to be used by other classes.ConstantsUtility

In this exercise, you will work with the class that stores several time constants for seconds. You need to add two static fields and and print them in this order, each of them on a new line.TimeConstantsSECONDS_IN_DAYSECONDS_IN_WEEK