Create Create an email
We need to create an email name for an employee. For this, implement a method that takes a String name and a String surname as an argument. The method should concatenate the name, the surname, and at the end, add the domain "@work.net". Return the result. It's guaranteed that input strings will not be empty.

Use StringBuilder to solve the problem.