Create Generate random numbers
You are given three numbers: n , a and b.

Your task is to output the sum of the first n random numbers in a range from a to b inclusively. The seed of the generator should be set as a + b.

The input contains numbers in a single line in the following order: n, a, b.